use vector rbegin

+++ gframe/network.h
@@ -200,7 +200,7 @@ inline bool check_msg_size(int size) {
 	return true;
 }
 
-inline unsigned int GetPosition(unsigned char* qbuf, int offset) {
+inline unsigned int GetPosition(unsigned char* qbuf, size_t offset) {
 	unsigned int info = 0;
 	std::memcpy(&info, qbuf + offset, sizeof info);
 	return info >> 24;
